backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.6",
    "img2pdf>=0.6.3",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27 ; sys_platform != 'win32'",
    "python-magic-bin>=0.4.14 ; sys_platform == 'win32'",
    "python-multipart>=0.0.22",
    "python-pptx>=0.6.23",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.40.0",
]
